Update to the new borrowing-based futures read/write API in gio
